Changes to tool parameters
Initial improvements to tool exposed outputs as parameters rather than hard coding an "add to map", this allows for model builder connectivity between tools. This update includes additional code for defining the _schema_ of the outputs so that tools can be fully utilised in model builder for automation.

Tool 3 had centreline parameter added to overcome a chicken and egg scenario where a request for the route ID from the route layer fails because in model builder when connecting tools the route layer does not actually exist until tool 2 has run. The centreline layer does exist and is the source of the route ID, hence the need to add it as a parameter to ensure correct continuity within a model.